Can use improvement but finally first blueprint and sharing at  the app level the non-serializable Device
